refactor: data type u64. Test add new data working

+++ src/main.rs
@@ -1,6 +1,6 @@
 use rusty_merkle_tree::MerkleTree;
 fn main() {
-    let data: Vec<u8> = vec![1, 2, 3, 4, 5,6];
+    let data: Vec<u64> = vec![1, 2, 3, 4, 5,6];
     let mut tree = MerkleTree::new(data);
     tree.complete_tree();
     println!("{:?}", tree);
